Added edit functionality to education component. Added unique educations for logged in users.

+++ bronco-resume-app/backend/routes/education.js
@@ -1,17 +1,19 @@
 const express = require('express');
-
+const checkAuth = require('../middleware/check-auth');
 const router = express.Router();
 
 const Education = require('../models/education');
 
-router.post('', (req, res, next) => {
+router.post('',
+  checkAuth, (req, res, next) => {
   const education = new Education({
     schoolName: req.body.schoolName,
     degreeType: req.body.degreeType,
     major: req.body.major,
     schoolStartDate: req.body.schoolStartDate,
     schoolEndDate: req.body.schoolEndDate,
-    gpa: req.body.gpa
+    gpa: req.body.gpa,
+    creator: req.userData.userId
   });
   education.save().then(createdEducation => {
     res.status(201).json({
@@ -21,8 +23,9 @@ router.post('', (req, res, next) => {
   });
 });
 
-router.get('', (req, res, next) => {
-  Education.find()
+router.get('',
+  checkAuth, (req, res, next) => {
+  Education.find({creator: req.userData.userId})
   .then(documents => {
     res.status(200).json({
       message: 'Education fetched successfully!',
@@ -31,8 +34,43 @@ router.get('', (req, res, next) => {
   })
 });
 
-router.delete('/:id', (req, res, next) => {
-  Education.deleteOne({ _id: req.params.id }).then(result => {
+router.get('/:id',
+  checkAuth, (req, res, next) => {
+  Education.findById(req.params.id).then(education => {
+    if (education) {
+      res.status(200).json(education);
+    } else {
+      res.status(404).json({message: 'Education not found!'});
+    }
+  })
+  .then(documents => {
+    res.status(200).json({
+      message: 'Education fetched successfully!',
+      education: documents
+    })
+  })
+});
+
+router.put('/:id',
+  checkAuth, (req, res, next) => {
+  const education = new Education({
+    schoolName: req.body.schoolName,
+    degreeType: req.body.degreeType,
+    major: req.body.major,
+    schoolStartDate: req.body.schoolStartDate,
+    schoolEndDate: req.body.schoolEndDate,
+    gpa: req.body.gpa,
+    _id: req.body.id
+  });
+  Education.updateOne({_id: req.params.id, creator: req.userData.userId}, education)
+    .then(result => {
+      res.status(200).json({message: "Education update successful!"})
+    })
+});
+
+router.delete('/:id',
+  checkAuth, (req, res, next) => {
+  Education.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
    console.log(result);
    res.status(200).json({ message: "Education deleted!" });
  });
